# apps/control-plane-executor/openapi/families/mcp.openapi.json
{
  "components": {
    "headers": {
      "XCorrelationId": {
        "description": "Correlation identifier resolved for this request. It is preserved when valid at ingress or generated when omitted.",
        "required": true,
        "schema": {
          "maxLength": 128,
          "minLength": 8,
          "pattern": "^[A-Za-z0-9._:-]{8,128}$",
          "type": "string"
        }
      }
    },
    "parameters": {
      "IdempotencyKey": {
        "description": "Stable idempotency token for safe retries of POST, PUT, PATCH, and DELETE operations routed through the gateway. Replays are retained for 24h and surfaced through X-Idempotency-Replayed.",
        "in": "header",
        "name": "Idempotency-Key",
        "required": true,
        "schema": {
          "maxLength": 128,
          "minLength": 8,
          "pattern": "^[A-Za-z0-9._:-]{8,128}$",
          "type": "string"
        }
      },
      "PageAfter": {
        "description": "Opaque cursor for the next page window.",
        "in": "query",
        "name": "page[after]",
        "required": false,
        "schema": {
          "maxLength": 256,
          "minLength": 1,
          "type": "string"
        }
      },
      "PageSize": {
        "description": "Maximum number of items to return for list and catalog routes.",
        "in": "query",
        "name": "page[size]",
        "required": false,
        "schema": {
          "default": 25,
          "maximum": 200,
          "minimum": 1,
          "type": "integer"
        }
      },
      "WorkspaceId": {
        "description": "Stable workspace identifier.",
        "in": "path",
        "name": "workspaceId",
        "required": true,
        "schema": {
          "minLength": 3,
          "pattern": "^wrk_[0-9a-z]+$",
          "type": "string"
        }
      },
      "XApiVersion": {
        "description": "Pinned additive contract version for the v1 public API release published with gateway QoS and hardening support.",
        "in": "header",
        "name": "X-API-Version",
        "required": true,
        "schema": {
          "const": "2026-03-26",
          "pattern": "^2026-03-26$",
          "type": "string"
        }
      },
      "XCorrelationId": {
        "description": "Optional caller correlation identifier. The gateway preserves one valid value or generates one when omitted, and every response returns the resolved value in X-Correlation-Id.",
        "in": "header",
        "name": "X-Correlation-Id",
        "required": false,
        "schema": {
          "maxLength": 128,
          "minLength": 8,
          "pattern": "^[A-Za-z0-9._:-]{8,128}$",
          "type": "string"
        }
      }
    },
    "schemas": {
      "ErrorDetail": {
        "additionalProperties": true,
        "properties": {
          "reason": {
            "type": "string"
          },
          "violations": {
            "items": {
              "type": "string"
            },
            "type": "array"
          }
        },
        "type": "object"
      },
      "ErrorResource": {
        "additionalProperties": false,
        "properties": {
          "id": {
            "type": "string"
          },
          "path": {
            "type": "string"
          },
          "type": {
            "type": "string"
          }
        },
        "required": [
          "path"
        ],
        "type": "object"
      },
      "ErrorResponse": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "pattern": "^GW_[A-Z0-9_]+$",
            "type": "string"
          },
          "correlationId": {
            "maxLength": 128,
            "minLength": 8,
            "pattern": "^[A-Za-z0-9._:-]{8,128}$",
            "type": "string"
          },
          "detail": {
            "$ref": "#/components/schemas/ErrorDetail"
          },
          "message": {
            "type": "string"
          },
          "requestId": {
            "maxLength": 128,
            "minLength": 8,
            "type": "string"
          },
          "resource": {
            "$ref": "#/components/schemas/ErrorResource"
          },
          "retryable": {
            "type": "boolean"
          },
          "status": {
            "maximum": 599,
            "minimum": 400,
            "type": "integer"
          },
          "timestamp": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "status",
          "code",
          "message",
          "detail",
          "requestId",
          "correlationId",
          "timestamp",
          "resource"
        ],
        "type": "object"
      },
      "GatewayMutationAccepted": {
        "additionalProperties": false,
        "properties": {
          "acceptedAt": {
            "format": "date-time",
            "type": "string"
          },
          "correlationId": {
            "maxLength": 128,
            "minLength": 8,
            "type": "string"
          },
          "family": {
            "type": "string"
          },
          "requestId": {
            "maxLength": 128,
            "minLength": 8,
            "type": "string"
          },
          "resourceId": {
            "type": "string"
          },
          "resourceType": {
            "type": "string"
          },
          "status": {
            "enum": [
              "accepted",
              "queued"
            ],
            "type": "string"
          }
        },
        "required": [
          "requestId",
          "status",
          "family",
          "resourceType",
          "resourceId",
          "correlationId",
          "acceptedAt"
        ],
        "type": "object"
      },
      "PageInfo": {
        "additionalProperties": false,
        "properties": {
          "after": {
            "type": "string"
          },
          "nextCursor": {
            "type": "string"
          },
          "size": {
            "maximum": 200,
            "minimum": 1,
            "type": "integer"
          }
        },
        "required": [
          "size"
        ],
        "type": "object"
      }
    }
  },
  "info": {
    "description": "Workspace-scoped MCP server publishing, version curation, and tool-call mediation through the control-plane executor. Generated from the unified public API contract for v1.",
    "summary": "Workspace-scoped MCP server publishing, version curation, and tool-call mediation through the control-plane executor.",
    "title": "In Falcone MCP API",
    "version": "1.21.0"
  },
  "jsonSchemaDialect": "https://json-schema.org/draft/2020-12/schema",
  "openapi": "3.1.0",
  "paths": {
    "/v1/mcp/workspaces/{workspaceId}/servers": {
      "get": {
        "operationId": "listMcpServers",
        "parameters": [
          {
            "$ref": "#/components/parameters/XApiVersion"
          },
          {
            "$ref": "#/components/parameters/XCorrelationId"
          },
          {
            "$ref": "#/components/parameters/WorkspaceId"
          },
          {
            "$ref": "#/components/parameters/PageSize"
          },
          {
            "$ref": "#/components/parameters/PageAfter"
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": false,
                  "properties": {
                    "items": {
                      "items": {
                        "additionalProperties": false,
                        "properties": {
                          "currentVersion": {
                            "type": "string"
                          },
                          "endpoint": {
                            "type": "string"
                          },
                          "id": {
                            "minLength": 1,
                            "type": "string"
                          },
                          "name": {
                            "minLength": 1,
                            "type": "string"
                          },
                          "status": {
                            "enum": [
                              "draft",
                              "pending_review",
                              "published",
                              "archived"
                            ],
                            "type": "string"
                          }
                        },
                        "required": [
                          "id",
                          "name",
                          "status"
                        ],
                        "type": "object"
                      },
                      "type": "array"
                    },
                    "pageInfo": {
                      "$ref": "#/components/schemas/PageInfo"
                    }
                  },
                  "required": [
                    "items"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "Workspace collection returned.",
            "headers": {
              "X-Correlation-Id": {
                "$ref": "#/components/headers/XCorrelationId"
              }
            }
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "The request is malformed or contextually inconsistent.",
            "headers": {
              "X-Correlation-Id": {
                "$ref": "#/components/headers/XCorrelationId"
              }
            }
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "The caller is authenticated but not authorized for this scope.",
            "headers": {
              "X-Correlation-Id": {
                "$ref": "#/components/headers/XCorrelationId"
              }
            }
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Gateway QoS policy rejected the request rate for this route family.",
            "headers": {
              "X-Correlation-Id": {
                "$ref": "#/components/headers/XCorrelationId"
              }
            }
          },
          "431": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Gateway rejected one or more invalid or oversized request headers.",
            "headers": {
              "X-Correlation-Id": {
                "$ref": "#/components/headers/XCorrelationId"
              }
            }
          },
          "504": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Gateway timed out while waiting for the downstream control-plane or adapter response.",
            "headers": {
              "X-Correlation-Id": {
                "$ref": "#/components/headers/XCorrelationId"
              }
            }
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unexpected control-plane error.",
            "headers": {
              "X-Correlation-Id": {
                "$ref": "#/components/headers/XCorrelationId"
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "List MCP servers for one workspace",
        "tags": [
          "MCP"
        ],
        "x-audiences": [
          "developer_external",
          "workspace_owner",
          "workspace_admin",
          "workspace_developer",
          "workspace_service_account",
          "platform_team"
        ],
        "x-family": "mcp",
        "x-owning-service": "control_api",
        "x-rate-limit-class": "data-read",
        "x-resource-type": "mcp_server",
        "x-scope": "workspace",
        "x-visibility": "public"
      },
      "post": {
        "operationId": "createMcpServer",
        "parameters": [
          {
            "$ref": "#/components/parameters/XApiVersion"
          },
          {
            "$ref": "#/components/parameters/XCorrelationId"
          },
          {
            "$ref": "#/components/parameters/WorkspaceId"
          },
          {
            "$ref": "#/components/parameters/IdempotencyKey"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": true,
                "properties": {
                  "description": {
                    "type": "string"
                  },
                  "name": {
                    "minLength": 1,
                    "type": "string"
                  },
                  "tools": {
                    "items": {
                      "additionalProperties": true,
                      "type": "object"
                    },
                    "type": "array"
                  }
                },
                "required": [
                  "name"
                ],
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GatewayMutationAccepted"
                }
              }
            },
            "description": "Workspace mutation accepted.",
            "headers": {
              "X-Correlation-Id": {
                "$ref": "#/components/headers/XCorrelationId"
              }
            }
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "The request is malformed or contextually inconsistent.",
            "headers": {
              "X-Correlation-Id": {
                "$ref": "#/components/headers/XCorrelationId"
              }
            }
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "The caller is authenticated but not authorized for this scope.",
            "headers": {
              "X-Correlation-Id": {
                "$ref": "#/components/headers/XCorrelationId"
              }
            }
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Gateway rejected the request body as too large for this route family.",
            "headers": {
              "X-Correlation-Id": {
                "$ref": "#/components/headers/XCorrelationId"
              }
            }
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Gateway QoS policy rejected the request rate for this route family.",
            "headers": {
              "X-Correlation-Id": {
                "$ref": "#/components/headers/XCorrelationId"
              }
            }
          },
          "431": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Gateway rejected one or more invalid or oversized request headers.",
            "headers": {
              "X-Correlation-Id": {
                "$ref": "#/components/headers/XCorrelationId"
              }
            }
          },
          "504": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Gateway timed out while waiting for the downstream control-plane or adapter response.",
            "headers": {
              "X-Correlation-Id": {
                "$ref": "#/components/headers/XCorrelationId"
              }
            }
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unexpected control-plane error.",
            "headers": {
              "X-Correlation-Id": {
                "$ref": "#/components/headers/XCorrelationId"
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Create an MCP server draft for one workspace",
        "tags": [
          "MCP"
        ],
        "x-audiences": [
          "developer_external",
          "workspace_owner",
          "workspace_admin",
          "workspace_developer",
          "workspace_service_account",
          "platform_team"
        ],
        "x-downstream-adapters": [
          "postgresql",
          "mcp-runtime"
        ],
        "x-family": "mcp",
        "x-owning-service": "control_api",
        "x-rate-limit-class": "data-write",
        "x-resource-type": "mcp_server",
        "x-scope": "workspace",
        "x-visibility": "public"
      }
    }
  },
  "servers": [
    {
      "url": "https://{publicHost}",
      "variables": {
        "publicHost": {
          "default": "api.dev.in-falcone.example.com",
          "enum": [
            "api.dev.in-falcone.example.com",
            "api.sandbox.in-falcone.example.com",
            "api.staging.in-falcone.example.com",
            "api.in-falcone.example.com"
          ]
        }
      }
    }
  ],
  "tags": [
    {
      "description": "MCP server hosting and tool-call mediation APIs.",
      "name": "MCP"
    }
  ]
}